change alt values from float to int

+++ runtime/include/AlterableValues.h
@@ -6,28 +6,28 @@ class AlterableValues
 {
 public:
 	AlterableValues() = default;
-	AlterableValues(const std::vector<float>& values) : Values(values) {}
+	AlterableValues(const std::vector<int>& values) : Values(values) {}
 
-	void SetValue(int index, float value) { 
+	void SetValue(int index, int value) { 
 		// Resize the vector if needed
 		if (index >= Values.size())
 			Values.resize(index + 1, 0);
 		Values[index] = value;
 	}
 
-	void AddValue(int index, float value) {
+	void AddValue(int index, int value) {
 		SetValue(index, GetValue(index) + value);
 	}
 	
-	void SubtractValue(int index, float value) {
+	void SubtractValue(int index, int value) {
 		SetValue(index, GetValue(index) - value);
 	}
 
-	float GetValue(int index) const {
+	int GetValue(int index) const {
 		if (index < 0 || index >= Values.size())
 			return 0;
 		return Values[index];
 	}
 private:
-	std::vector<float> Values;
+	std::vector<int> Values;
 };
